Update 24/06: Expansion Chip System added!!!!
Feature: Expansion Chip System.
Usage:
 - Press R to open the Expansion Chip Menu. Press E to open the Inventory Menu.
 - Chips are locked and unusable by default. Collect them at the scene as Drop Item to unlock
 - Press button to use it!
TODO:
 - Implement Chip Load Limitation
 - Add more chip buffs to the game (and also the Icons too.
*How to add chip buff*
 - Create an Expansion Chip Scriptable Object in the Editor
 - Open the Player UI Prefab, click on an empty slot, drag the new SO into it.
 - Create a script for the buff/debuff, using the interface IChip. Receive the equip state from the Expansion Chip Slot and apply/remove effect.

+++ Assets/Scripts/GamePlay/Combat/DMG Calculation/DamageOutCalculation.cs
@@ -6,16 +6,16 @@ using UnityEngine;
 public class DamageOutCalculator : MonoBehaviour
 {
     //The total amount of damage boost (uses %)
-    private int totalDMGBoost = 0;
-    public int _totalDMGBoost { get { return totalDMGBoost; } }
+    private float totalDMGBoost = 0;
+    public float _totalDMGBoost { get { return totalDMGBoost; } }
     
     //Increase the totalDMGBoost by the amount of the boost
-    public void IncreaseDMGBoost(int boost)
+    public void IncreaseDMGBoost(float boost)
     {
         totalDMGBoost += boost;
     }
     //Decrease the totalDMGBoost by the amount of the boost
-    public void DecreaseDMGBoost(int boost)
+    public void DecreaseDMGBoost(float boost)
     {
         totalDMGBoost -= boost;
     }
